requirements.txt
astroid==1.5.3
certifi==2016.2.28
Geometry2D==0.7.0
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
Mesa==0.8.0
numpy==1.13.1
pandas==0.20.3
pygame==1.9.3
pylint==1.7.2
python-dateutil==2.6.1
pytz==2017.2
RenderingFacilities==0.1.0
six==1.11.0
tornado==4.5.2
tqdm==4.17.1
wrapt==1.10.11
